doit: Split out cmd_devmem
