+++ cmd/devmem.c
@@ -1,0 +1,48 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (C) 2018,2021 IBM Corp.
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+#include "ahb.h"
+#include "ast.h"
+#include "devmem.h"
+#include "priv.h"
+
+int cmd_devmem(const char *name, int argc, char *argv[])
+{
+    struct devmem _devmem, *devmem = &_devmem;
+    struct ahb _ahb, *ahb = &_ahb;
+    int cleanup;
+    int rc;
+
+    rc = devmem_init(devmem);
+    if (rc < 0) {
+        bool denied = (rc == -EACCES || rc == -EPERM);
+        if (denied && !priv_am_root()) {
+            priv_print_unprivileged(name);
+        } else {
+            errno = -rc;
+            perror("devmem_init");
+        }
+        exit(EXIT_FAILURE);
+    }
+
+    ahb_use(ahb, ahb_devmem, devmem);
+    rc = ast_ahb_access(name, argc, argv, ahb);
+    if (rc) {
+        errno = -rc;
+        perror("ast_ahb_access");
+        exit(EXIT_FAILURE);
+    }
+
+    cleanup = devmem_destroy(devmem);
+    if (cleanup) {
+        errno = -cleanup;
+        perror("devmem_destroy");
+        exit(EXIT_FAILURE);
+    }
+
+    return 0;
+}
+
